+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to
 - 💬(wording) error page wording #102
 - ⚡️(web-search) allow to override returned chunks #107
 - 🐛(activation-codes) create contact in brevo before add to list #108
+- ⚗️(summarization) add system prompt to handle tool #112
 
 
 ## [0.0.1] - 2025-10-19